+++ test/forms/Validators.test.js
@@ -175,12 +175,36 @@ describe('Validators', () => {
             expect(isNumber(100)).to.be.true;
         });
 
+        it('should return true when the value is the string "100"', () => {
+            expect(isNumber('100')).to.be.true;
+        });
+
         it('should return true when the value is 1.9', () => {
             expect(isNumber(1.9)).to.be.true;
         });
 
+        it('should return true when the value is the string "1.9"', () => {
+            expect(isNumber('1.9')).to.be.true;
+        });
+
+        it('should return false when the value is an arbitrary string that starts with a number', () => {
+            expect(isNumber('13 bananas')).to.be.false;
+        });
+
+        it('should return true when the value is a string containing a number in scientific notation', () => {
+            expect(isNumber('314159e-5')).to.be.true;
+        });
+
+        it('should return true when the value is a number in scientific notation', () => {
+            expect(isNumber(1.234e-45)).to.be.true;
+        });
+
         it('should return true when the value is 0', () => {
             expect(isNumber(0)).to.be.true;
+        });
+
+        it('should return true when the value is the string "0"', () => {
+            expect(isNumber('0')).to.be.true;
         });
 
         it('should return false when the value is Infinity', () => {
@@ -195,8 +219,16 @@ describe('Validators', () => {
             expect(isNumber(new Number(2.1))).to.be.true; // eslint-disable-line no-new-wrappers
         });
 
-        it('should return false when the object.toString returns a number', () => {
-            expect(isNumber({toString() { return ''; }})).to.be.false;
+        it('should return true when object.toString() returns the empty string', () => {
+            expect(isNumber({toString() { return ''; }})).to.be.true;
+        });
+
+        it('should return false when object.toString() returns an arbitrary string', () => {
+            expect(isNumber({toString() { return 'bla'; }})).to.be.false;
+        });
+
+        it('should return true when object.toString() returns a numeric string', () => {
+            expect(isNumber({toString() { return '1'; }})).to.be.true;
         });
 
         it('should return true if the value is undefined', () => {