+++ CHANGELOG.md
@@ -4,6 +4,36 @@ D2-UI Changelog
 HEAD
 ----
 
-- Added FormBuilder component
-- Added CHANGELOG.md
+**Breaking changes**
+
+- `Validators.isNumber()` now accepts numbers, numeric strings, the empty string and
+  `undefined` (see #34)
+
+
+
+0.0.17 (2016-03-04)
+-------------------
+
+- [Minor] Upgrade dependencies
+
+
+
+0.0.16 (2016-02-26)
+-------------------
+
+- [Minor] Upgrade to React 0.14
+
+
+
+0.0.15 (2016-02-24)
+-------------------
+
+- [Minor] Code style
+
+
+
+0.0.14 (2016-02-18)
+-------------------
+- [FormBuilder] Add FormBuilder component
+- Add CHANGELOG.md